requirements.txt
absl-py==2.1.0
docker==7.1.0
Mako==1.3.5
